// src/services/sync/protocol/message-types.ts
export enum MessageType {
  CLIENT_REQUEST_SYNC = 'CLIENT_REQUEST_SYNC',
  SERVER_ACK_SYNC_REQUEST = 'SERVER_ACK_SYNC_REQUEST',

  CLIENT_SEND_LOCKED_PATIENT_IDS = 'CLIENT_SEND_LOCKED_PATIENT_IDS',
  SERVER_SEND_UPDATED_PATIENTS = 'SERVER_SEND_UPDATED_PATIENTS',

  CLIENT_SEND_NEW_PATIENTS = 'CLIENT_SEND_NEW_PATIENTS',
  SERVER_ACK_NEW_PATIENTS = 'SERVER_ACK_NEW_PATIENTS',

  SERVER_REQUEST_PATIENT_MEASURES = 'SERVER_REQUEST_PATIENT_MEASURES',
  CLIENT_SEND_PATIENT_MEASURES = 'CLIENT_SEND_PATIENT_MEASURES',
  SERVER_ACK_PATIENT_MEASURES = 'SERVER_ACK_PATIENT_MEASURES',

  SERVER_SEND_PRO_PATIENTS = 'SERVER_SEND_PRO_PATIENTS',
  CLIENT_ACK_PRO_PATIENTS = 'CLIENT_ACK_PRO_PATIENTS',

  SERVER_REQUEST_CACHE_MANIFEST = 'SERVER_REQUEST_CACHE_MANIFEST',
  CLIENT_SEND_CACHE_MANIFEST = 'CLIENT_SEND_CACHE_MANIFEST',
  SERVER_SEND_MISSING_REFERENCES = 'SERVER_SEND_MISSING_REFERENCES',
  CLIENT_ACK_REFERENCES = 'CLIENT_ACK_REFERENCES',

  SERVER_SEND_ACTIVE_TASKS = 'SERVER_SEND_ACTIVE_TASKS',
  CLIENT_ACK_ACTIVE_TASKS = 'CLIENT_ACK_ACTIVE_TASKS',

  SERVER_REQUEST_TASK_RESULTS = 'SERVER_REQUEST_TASK_RESULTS',
  CLIENT_SEND_TASK_RESULTS = 'CLIENT_SEND_TASK_RESULTS',
  SERVER_ACK_TASK_RESULTS = 'SERVER_ACK_TASK_RESULTS',

  SERVER_CLOSE_SYNC_SESSION = 'SERVER_CLOSE_SYNC_SESSION',
  ERROR = 'ERROR',
  CLIENT_DISCONNECTED = 'CLIENT_DISCONNECTED',
}